+++ data/fileio/CSVFormat.h
@@ -86,8 +86,8 @@ public:
     ModelType    getModelType()     const { return m_modelType;     }
     TimingType   getTimingType()    const { return m_timingType;    }
     TimeUnits    getTimeUnits()     const { return m_timeUnits;     }
-    size_t       getSampleRate()    const { return m_sampleRate;    }
-    size_t       getWindowSize()    const { return m_windowSize;    }
+    int          getSampleRate()    const { return m_sampleRate;    }
+    int          getWindowSize()    const { return m_windowSize;    }
     int          getColumnCount()   const { return m_columnCount;   }
     bool         getAllowQuoting()  const { return m_allowQuoting;  }
     QChar        getSeparator()     const { 
@@ -99,8 +99,8 @@ public:
     void setTimingType(TimingType t)      { m_timingType   = t; }
     void setTimeUnits(TimeUnits t)        { m_timeUnits    = t; }
     void setSeparator(QChar s)            { m_separator    = s; }
-    void setSampleRate(size_t r)          { m_sampleRate   = r; }
-    void setWindowSize(size_t s)          { m_windowSize   = s; }
+    void setSampleRate(int r)          { m_sampleRate   = r; }
+    void setWindowSize(int s)          { m_windowSize   = s; }
     void setColumnCount(int c)            { m_columnCount  = c; }
     void setAllowQuoting(bool q)          { m_allowQuoting = q; }
 
@@ -123,8 +123,8 @@ protected:
     TimingType   m_timingType;
     TimeUnits    m_timeUnits;
     QString      m_separator;
-    size_t       m_sampleRate;
-    size_t       m_windowSize;
+    int          m_sampleRate;
+    int          m_windowSize;
 
     int          m_columnCount;
     bool         m_variableColumnCount;
